integrity/requirements.txt
pyelftools==0.27